requirements.txt
redis
msgraph-sdk
sanic
azure-identity
pyyaml
